improved material handling, only ad 1 chest to database (doublechest)

+++ src/main/java/de/maniacraft/chestindex/listeners/ChestBlockListener.java
@@ -3,6 +3,7 @@ package de.maniacraft.chestindex.listeners;
 import org.bukkit.Material;
 import org.bukkit.World;
 import org.bukkit.block.Block;
+import org.bukkit.block.BlockFace;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
@@ -24,11 +25,34 @@ public class ChestBlockListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL)
 	public void onBlockBreak(BlockBreakEvent event) {
+		if (event.isCancelled()) {
+			return;
+		}
 		if (!Chestindex.LWC) {
 			Block block = event.getBlock();
+			Block block2 = null;
 			World world = block.getWorld();
-			if (plugin.DB.isListed(world, block.getX(), block.getY(), block.getZ())) {
-				plugin.DB.removeChest(world, block.getX(), block.getY(), block.getZ());
+			if (block.getType() == Material.CHEST) {
+				BlockFace[] faces = new BlockFace[] { BlockFace.NORTH, BlockFace.SOUTH, BlockFace.EAST, BlockFace.WEST };
+
+				// Double Chest
+				for (BlockFace blockFace : faces) {
+					Block face = block.getRelative(blockFace);
+					if (face.getType() == Material.CHEST) {
+						block2 = face;
+					}
+				}
+				// Remove or edit Chestentry
+				if (plugin.DB.isListed(world, block.getX(), block.getY(), block.getZ())) {
+					// Doppelchest, Chest die in der Datenbank ist wird geloescht -> Trage zweite Chest stattdessen ein.
+					if (block2 != null && !plugin.DB.isListed(world, block2.getX(), block2.getY(), block2.getZ())) {
+						plugin.DB.editChest(world, block.getX(), block.getY(), block.getZ(), block2.getX(), block2.getY(), block2.getZ());
+					}
+					// Eizelne Kiste die in der Datenbank ist wird geloescht -> Loesche aus Datenbank.
+					else {
+						plugin.DB.removeChest(world, block.getX(), block.getY(), block.getZ());
+					}
+				}
 			}
 		}
 	}
@@ -39,11 +63,26 @@ public class ChestBlockListener implements Listener {
 			return;
 		}
 		if (!Chestindex.LWC) {
+
 			Player player = event.getPlayer();
 			Block block = event.getBlockPlaced();
+			Block block2 = null;
 			World world = block.getWorld();
 			if (block.getType() == Material.CHEST) {
-				plugin.DB.addChest(player, world, block.getX(), block.getY(), block.getZ());
+				BlockFace[] faces = new BlockFace[] { BlockFace.NORTH, BlockFace.SOUTH, BlockFace.EAST, BlockFace.WEST };
+
+				// Double Chest
+				for (BlockFace blockFace : faces) {
+					Block face = block.getRelative(blockFace);
+					if (face.getType() == Material.CHEST) {
+						block2 = face;
+					}
+				}
+				
+				// Entweder Doppelchest, aber Chest2 ist nicht gelistet, oder Einzelchest -> trage ein
+				if (block2 == null || !plugin.DB.isListed(world, block2.getX(), block2.getY(), block2.getZ())) {
+					plugin.DB.addChest(player, world, block.getX(), block.getY(), block.getZ());
+				}
 			}
 		}
 	}
